Apply requested changes to swing and belts commands
- Change parameters from double to float
- Add validation that fixed belt and moving belt must not be the same
- Change "Moving belt hit current limit" from log_error to log_warn
- Add comments emphasizing belts must be in comply mode before other belts move
- Update sscanf format specifiers from %lf to %f
- Update strtod to strtof for parsing float values

+++ FluidNC/src/Maslow/Calibration.h
@@ -72,7 +72,7 @@ public:
     void setSafety(bool state);
     
     // Advanced belt control commands
-    bool swing(const char* fixedBeltName, const char* movingBeltName, double distance, double speed);
+    bool swing(const char* fixedBeltName, const char* movingBeltName, float distance, float speed);
     bool belts(const char* value);
 
     //State machine functions
